(svn r14494) -Fix: few MSVC x64 warnings.

+++ src/gfx.cpp
@@ -288,12 +288,12 @@ static void HandleBiDiAndArabicShapes(char *buffer, const char *lastof)
 	input_output[length] = 0;
 
 	UErrorCode err = U_ZERO_ERROR;
-	UBiDi *para = ubidi_openSized(length, 0, &err);
+	UBiDi *para = ubidi_openSized((int32_t)length, 0, &err);
 	if (para == NULL) return;
 
-	ubidi_setPara(para, input_output, length, UBIDI_DEFAULT_RTL, NULL, &err);
-	ubidi_writeReordered(para, intermediate, length, 0, &err);
-	length = u_shapeArabic(intermediate, length, input_output, lengthof(input_output), U_SHAPE_TEXT_DIRECTION_VISUAL_LTR | U_SHAPE_LETTERS_SHAPE, &err);
+	ubidi_setPara(para, input_output, (int32_t)length, UBIDI_DEFAULT_RTL, NULL, &err);
+	ubidi_writeReordered(para, intermediate, (int32_t)length, 0, &err);
+	length = u_shapeArabic(intermediate, (int32_t)length, input_output, lengthof(input_output), U_SHAPE_TEXT_DIRECTION_VISUAL_LTR | U_SHAPE_LETTERS_SHAPE, &err);
 	ubidi_close(para);
 
 	if (U_FAILURE(err)) return;
